moved abstract-view into templates
